Добавление изменений
